Application ready. Edition 1.1
