+++ cookapp.http
@@ -52,24 +52,25 @@ Content-Type: application/json
     "5) Посыпать кунжутными семечками и заправить майонезом."
   ]
 }
+### ИНГРЕДИЕНТЫ -----------------------------------------------------------------------------------
 
 ### Получить ингредиент по id.
-GET http://localhost:8080/cookapp/ing/2
+GET http://localhost:8080/ingredient/2
 
 ### Создать ингредиент 1. id 0.
-POST http://localhost:8080/cookapp/add
+POST http://localhost:8080/ingredient/add
 Content-Type: application/json
 
 { "ingredientName": "Помидоры", "count": 2, "unit": "шт." }
 
 ### Создать ингредиент 2. id 1.
-POST http://localhost:8080/cookapp/add
+POST http://localhost:8080/ingredient/add
 Content-Type: application/json
 
 { "ingredientName": "Королевские креветки", "count": 300, "unit": "г." }
 
 ### Создать ингредиент 3. id 2.
-POST http://localhost:8080/cookapp/add
+POST http://localhost:8080/ingredient/add
 Content-Type: application/json
 
 { "ingredientName": "Зеленый салат", "count": 1, "unit": "пучок" }
