+++ src/main/java/com/example/cookapp/service/CookAppService.java
@@ -12,13 +12,10 @@ import java.util.Map;
 @Service
 public class CookAppService {
     private final Map<Integer, Recipe> recipeMap = new HashMap<>();
-    private final Map<Integer, Ingredient> ingredientMap = new HashMap<>();
+    private final Map<Integer, Ingredient> ingredientMap = new HashMap<>(); // переносим в IngredientService
 
     public Recipe addRecipe(Recipe recipe) {
         return recipeMap.put(recipeMap.size(), recipe);
-    }
-    public Ingredient addIngredient(Ingredient ingredient) {
-        return ingredientMap.put(ingredientMap.size(), ingredient);
     }
     public Recipe getTheRecipe(Integer id) {
         if (recipeMap.containsKey(id)) {
@@ -26,13 +23,18 @@ public class CookAppService {
         } else
             throw new RuntimeException("Pецепта с таким id нет.");
     }
-    public Ingredient getTheIngredient(Integer id) {
-        if (ingredientMap.containsKey(id)) {
-            return ingredientMap.get(id);
-        } else
-            throw new RuntimeException("Pецепта с таким id нет.");
-    }
     public Collection<Recipe> getAllRecipe() {
         return recipeMap.values();
     }
+    // Ingredients...........................................................................
+//    public Ingredient getTheIngredient(Integer id) { // переносим в IngredientService
+//        if (ingredientMap.containsKey(id)) {
+//            return ingredientMap.get(id);
+//        } else
+//            throw new RuntimeException("Pецепта с таким id нет.");
+//    }
+//    public Ingredient addIngredient(Ingredient ingredient) {  // переносим в IngredientService
+//        return ingredientMap.put(ingredientMap.size(), ingredient);
+//    }
+
 }
